+++ ios/Classes/Module/SignalingManager.swift
@@ -14,6 +14,12 @@ public class SignalingManager: BaseServiceManager {
         self["signalingHungUp"] = signalingHungUp
         self["signalingGetRoomByGroupID"] = signalingGetRoomByGroupID
         self["signalingGetTokenByRoomID"] = signalingGetTokenByRoomID
+        self["signalingUpdateMeetingInfo"] = signalingUpdateMeetingInfo
+        self["signalingCreateMeeting"] = signalingCreateMeeting
+        self["signalingJoinMeeting"] = signalingJoinMeeting
+        self["signalingOperateStream"] = signalingOperateStream
+        self["signalingGetMeetings"] = signalingGetMeetings
+        self["signalingCloseRoom"] = signalingCloseRoom
     }
 
     func setSignalingListener(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
@@ -52,9 +58,33 @@ public class SignalingManager: BaseServiceManager {
     func signalingGetTokenByRoomID(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
         Open_im_sdkSignalingGetTokenByRoomID(BaseCallback(result: result), methodCall[string: "operationID"], methodCall[string: "roomID"])
     }
+    
+    func signalingUpdateMeetingInfo(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
+        Open_im_sdkSignalingUpdateMeetingInfo(BaseCallback(result: result), methodCall[string: "operationID"], methodCall[jsonString: "info"])
+    }
+    
+    func signalingCreateMeeting(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
+        Open_im_sdkSignalingCreateMeeting(BaseCallback(result: result), methodCall[string: "operationID"], methodCall[jsonString: "info"])
+    }
+    
+    func signalingJoinMeeting(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
+        Open_im_sdkSignalingJoinMeeting(BaseCallback(result: result), methodCall[string: "operationID"], methodCall[jsonString: "info"])
+    }
+    
+    func signalingOperateStream(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
+        Open_im_sdkSignalingOperateStream(BaseCallback(result: result), methodCall[string: "operationID"], methodCall[string: "streamType"], methodCall[string: "roomID"], methodCall[string: "userID"], methodCall[bool: "mute"], methodCall[bool: "muteAll"])
+    }
+    
+    func signalingGetMeetings(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
+        Open_im_sdkSignalingGetMeetings(BaseCallback(result: result), methodCall[string: "operationID"])
+    }
+    
+    func signalingCloseRoom(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
+        Open_im_sdkSignalingCloseRoom(BaseCallback(result: result), methodCall[string: "operationID"], methodCall[string: "roomID"])
+    }
 }
 public class SignalingListener: NSObject, Open_im_sdk_callbackOnSignalingListenerProtocol {
-   
+  
     private let channel:FlutterMethodChannel
     
     init(channel:FlutterMethodChannel) {
@@ -100,4 +130,9 @@ public class SignalingListener: NSObject, Open_im_sdk_callbackOnSignalingListene
     public func onRoomParticipantDisconnected(_ s: String?) {
         CommonUtil.emitEvent(channel: channel, method: "signalingListener", type: "onRoomParticipantDisconnected", errCode: nil, errMsg: nil, data: s)
     }
+    
+    public func onStreamChange(_ s: String?) {
+        CommonUtil.emitEvent(channel: channel, method: "signalingListener", type: "onStreamChange", errCode: nil, errMsg: nil, data: s)
+    }
+    
 }