+++ lib/src/listener/signaling_listener.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_openim_sdk/src/models/signaling_info.dart';
+import 'package:flutter_openim_sdk/flutter_openim_sdk.dart';
 
 /// 信令监听
 class OnSignalingListener {
@@ -12,6 +12,7 @@ class OnSignalingListener {
   final Function(SignalingInfo info)? onHangup;
   final Function(RoomCallingInfo info)? onRoomParticipantConnected;
   final Function(RoomCallingInfo info)? onRoomParticipantDisconnected;
+  final Function(MeetingStreamEvent event)? onMeetingStreamChanged;
 
   OnSignalingListener({
     this.onInvitationCancelled,
@@ -24,6 +25,7 @@ class OnSignalingListener {
     this.onHangup,
     this.onRoomParticipantConnected,
     this.onRoomParticipantDisconnected,
+    this.onMeetingStreamChanged,
   });
 
   /// 被邀请者收到：邀请者取消音视频通话
@@ -72,5 +74,9 @@ class OnSignalingListener {
 
   void roomParticipantDisconnected(RoomCallingInfo info) {
     onRoomParticipantDisconnected?.call(info);
+  }
+
+  void streamChangedEvent(MeetingStreamEvent event) {
+    onMeetingStreamChanged?.call(event);
   }
 }